backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
sqlalchemy==2.0.25
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
websockets==12.0
httpx==0.26.0
requests==2.31.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]
